api/requirements.txt

fastapi==0.115.4
uvicorn==0.32.0
pydantic==2.9.2
python-multipart==0.0.9
Pillow==10.4.0
